# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LeadForge Elite API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: leads
    description: Lead management
  - name: crm
    description: CRM pipeline operations
  - name: outreach
    description: Outreach campaigns
  - name: dashboard
    description: Dashboard analytics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads with filtering
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: score
          in: query
          schema:
            type: string
            enum: [Hot, Warm, Cold]
        - name: industry
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadListResponse"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"

  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    put:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Lead deleted

  /leads/{id}/notes:
    get:
      operationId: getLeadNotes
      tags: [leads]
      summary: Get notes for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: addLeadNote
      tags: [leads]
      summary: Add note to a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  /crm/pipeline:
    get:
      operationId: getCrmPipeline
      tags: [crm]
      summary: Get all leads grouped by CRM pipeline stage
      responses:
        "200":
          description: Pipeline stages with leads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineResponse"

  /crm/leads/{id}/stage:
    put:
      operationId: updateLeadStage
      tags: [crm]
      summary: Update lead CRM stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStageBody"
      responses:
        "200":
          description: Stage updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"

  /outreach/campaigns:
    get:
      operationId: listCampaigns
      tags: [outreach]
      summary: List outreach campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [outreach]
      summary: Create outreach campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  /outreach/campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [outreach]
      summary: Get campaign details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  /outreach/templates:
    get:
      operationId: listTemplates
      tags: [outreach]
      summary: List message templates
      responses:
        "200":
          description: Template list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [outreach]
      summary: Create message template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateBody"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/lead-scores:
    get:
      operationId: getLeadScoreBreakdown
      tags: [dashboard]
      summary: Get lead count grouped by score category
      responses:
        "200":
          description: Score breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoreBreakdown"

  /dashboard/pipeline-values:
    get:
      operationId: getPipelineValues
      tags: [dashboard]
      summary: Get lead counts per CRM stage
      responses:
        "200":
          description: Pipeline breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineStageCount"

  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      responses:
        "200":
          description: Recent activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /dashboard/industry-breakdown:
    get:
      operationId: getIndustryBreakdown
      tags: [dashboard]
      summary: Get leads grouped by industry
      responses:
        "200":
          description: Industry breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IndustryCount"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Lead:
      type: object
      properties:
        id:
          type: integer
        businessName:
          type: string
        ownerName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        whatsappActive:
          type: boolean
          nullable: true
        website:
          type: string
          nullable: true
        linkedinProfile:
          type: string
          nullable: true
        companySize:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        techStack:
          type: string
          nullable: true
        seoScore:
          type: integer
          nullable: true
        performanceScore:
          type: integer
          nullable: true
        leadScore:
          type: integer
        scoreCategory:
          type: string
          enum: [Hot, Warm, Cold]
        crmStage:
          type: string
          enum: [New Lead, Contacted, Qualified, Proposal Sent, Closed Won, Closed Lost]
        tags:
          type: array
          items:
            type: string
        noWebsite:
          type: boolean
        poorSeo:
          type: boolean
        noSocialPresence:
          type: boolean
        mobileUnfriendly:
          type: boolean
        aiInsight:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - businessName
        - leadScore
        - scoreCategory
        - crmStage
        - tags
        - noWebsite
        - poorSeo
        - noSocialPresence
        - mobileUnfriendly
        - createdAt
        - updatedAt

    CreateLeadBody:
      type: object
      properties:
        businessName:
          type: string
        ownerName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        whatsappActive:
          type: boolean
          nullable: true
        website:
          type: string
          nullable: true
        linkedinProfile:
          type: string
          nullable: true
        companySize:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        techStack:
          type: string
          nullable: true
        seoScore:
          type: integer
          nullable: true
        performanceScore:
          type: integer
          nullable: true
        noWebsite:
          type: boolean
        poorSeo:
          type: boolean
        noSocialPresence:
          type: boolean
        mobileUnfriendly:
          type: boolean
        tags:
          type: array
          items:
            type: string
        crmStage:
          type: string
          enum: [New Lead, Contacted, Qualified, Proposal Sent, Closed Won, Closed Lost]
      required: [businessName]

    UpdateLeadBody:
      type: object
      properties:
        businessName:
          type: string
        ownerName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        whatsappActive:
          type: boolean
          nullable: true
        website:
          type: string
          nullable: true
        linkedinProfile:
          type: string
          nullable: true
        companySize:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        techStack:
          type: string
          nullable: true
        seoScore:
          type: integer
          nullable: true
        performanceScore:
          type: integer
          nullable: true
        noWebsite:
          type: boolean
        poorSeo:
          type: boolean
        noSocialPresence:
          type: boolean
        mobileUnfriendly:
          type: boolean
        tags:
          type: array
          items:
            type: string
        crmStage:
          type: string
          enum: [New Lead, Contacted, Qualified, Proposal Sent, Closed Won, Closed Lost]
        aiInsight:
          type: string
          nullable: true

    Note:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
      required: [id, leadId, content, createdAt]

    CreateNoteBody:
      type: object
      properties:
        content:
          type: string
      required: [content]

    PipelineResponse:
      type: object
      properties:
        stages:
          type: array
          items:
            $ref: "#/components/schemas/PipelineStage"
      required: [stages]

    PipelineStage:
      type: object
      properties:
        stage:
          type: string
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
      required: [stage, leads]

    UpdateStageBody:
      type: object
      properties:
        stage:
          type: string
          enum: [New Lead, Contacted, Qualified, Proposal Sent, Closed Won, Closed Lost]
      required: [stage]

    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        channel:
          type: string
          enum: [Email, WhatsApp, LinkedIn]
        status:
          type: string
          enum: [Draft, Active, Paused, Completed]
        templateId:
          type: integer
          nullable: true
        leadCount:
          type: integer
        sentCount:
          type: integer
        openCount:
          type: integer
        replyCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, channel, status, leadCount, sentCount, openCount, replyCount, createdAt]

    CreateCampaignBody:
      type: object
      properties:
        name:
          type: string
        channel:
          type: string
          enum: [Email, WhatsApp, LinkedIn]
        templateId:
          type: integer
          nullable: true
      required: [name, channel]

    Template:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        channel:
          type: string
          enum: [Email, WhatsApp, LinkedIn]
        subject:
          type: string
          nullable: true
        body:
          type: string
        createdAt:
          type: string
      required: [id, name, channel, body, createdAt]

    CreateTemplateBody:
      type: object
      properties:
        name:
          type: string
        channel:
          type: string
          enum: [Email, WhatsApp, LinkedIn]
        subject:
          type: string
          nullable: true
        body:
          type: string
      required: [name, channel, body]

    DashboardStats:
      type: object
      properties:
        totalLeads:
          type: integer
        hotLeads:
          type: integer
        warmLeads:
          type: integer
        coldLeads:
          type: integer
        contactedLeads:
          type: integer
        qualifiedLeads:
          type: integer
        closedWon:
          type: integer
        activeCampaigns:
          type: integer
        emailsSentThisMonth:
          type: integer
        avgLeadScore:
          type: number
      required:
        - totalLeads
        - hotLeads
        - warmLeads
        - coldLeads
        - contactedLeads
        - qualifiedLeads
        - closedWon
        - activeCampaigns
        - emailsSentThisMonth
        - avgLeadScore

    ScoreBreakdown:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required: [category, count, percentage]

    PipelineStageCount:
      type: object
      properties:
        stage:
          type: string
        count:
          type: integer
      required: [stage, count]

    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        leadName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, type, description, createdAt]

    IndustryCount:
      type: object
      properties:
        industry:
          type: string
        count:
          type: integer
      required: [industry, count]

    LeadListResponse:
      type: object
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [leads, total, page, limit]